adding Quran with English and updates to files

+++ flask_app.py
@@ -47,40 +47,46 @@ class Verse(db.Model):
     nVerse = db.Column(db.Integer)
     verse = db.Column(db.String(4096))
 
-
+# Variables to determine whether the raw Quran input should be loaded upon running
+#   the script, and if it should be loaded, just a test portion or the whole Quran
+quran_load = True
+quran_test = True
 
 # Load Quran into table. First delete it, then add rows.
-n = Verse.query.delete()
-db.session.commit()
 
-quran_file = "/home/navid/mysite/quran-simple.txt"
+if quran_load:
+    n = Verse.query.delete()
+    db.session.commit()
 
-testlines=[]
-with open(quran_file, "r", encoding="utf-8") as f:
-    lines = [l.strip() for l in f.readlines()]
-    for x in range(10):
-        testlines.append(lines[x])
+    quran_file = "/home/navid/mysite/eng_quran_out.txt"
 
-test = True
+    testlines=[]
+    with open(quran_file, "r", encoding="utf-8") as f:
+        lines = [l.strip() for l in f.readlines()]
+        for x in range(10):
+            testlines.append(lines[x])
 
-if test:
-    n = 111
-else:
-    n = 50100
 
-for l in lines[0:n]:
-    # print(l)
-    temp_verse = verse2dict(l)
-    # print(temp_verse)
-    temp_verse = Verse(nSura=temp_verse["sura"],
-                      nVerse=temp_verse["verse"],
-                      verse=temp_verse["text"])
 
-    db.session.add(temp_verse)
-db.session.commit()
+    if quran_test:
+        n = 111
+    else:
+        n = 50100
 
-# testlines = Verse.query.limit(10).all()
+    for l in lines[0:n]:
+        # print(l)
+        temp_verse = verse2dict(l)
+        # print(temp_verse)
+        temp_verse = Verse(nSura=temp_verse["sura"],
+                          nVerse=temp_verse["verse"],
+                          verse=temp_verse["text"])
 
+        db.session.add(temp_verse)
+    db.session.commit()
+
+
+
+# Store queried verses
 verse_obj = []
 
 # Main web application function
@@ -107,7 +113,8 @@ def index():
             db.session.commit()
         elif execute_this == "Get verse":
             req = request.form["contents"].strip()
-
+            if not re.match("\d+\D+\d+\D*",req):
+                pass
             verse_obj = query_verses(req, db)
         return redirect(url_for('index'))
 
@@ -124,8 +131,9 @@ def wibble():
 # process request to query
 def query_verses(req, db):
     req = re.split("\D+", req)
+    print(req)
     if len(req) < 3:
-        req[2], req[3] = req[0], req[1]
+        req.extend( [ req[0], req[1] ] )
 
     begSura = req[0]
     begVerse = req[1]
@@ -137,7 +145,7 @@ def query_verses(req, db):
         rv = db.session.execute(textual, {"x1": begSura, "x2": begVerse, "x4": endVerse}).fetchall()
     else:
         textual = text("select nSura, nVerse, verse from q_tbl where (nSura = :x1 and nVerse >= :x2)" +
-        " or (nSura > :x1 and nSura < :x3) or (nSura = :x3 and nVerse <= :x4")
+        " or (nSura > :x1 and nSura < :x3) or (nSura = :x3 and nVerse <= :x4)")
         rv = db.session.execute(textual, {"x1": begSura, "x2": begVerse, "x3": endSura, "x4": endVerse}).fetchall()
 
     return load_query_into_dict(rv)
@@ -151,7 +159,7 @@ def load_query_into_dict(queried_verses):
         di = d.copy()
         di["nSura"] = v[0]
         di["nVerse"] = v[1]
-        di["text"] = v[2]
+        di["verse"] = v[2]
         rv.append(di.copy())
 
     return rv
